requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx==0.26.0
